Moved src/ to lib/
