+++ lib/loghose.js
